Print metrics exported by PrometheusExporter
To make it easier to keep the exposed metrics up to date, I've added a
helper to the prometheus exporter. It discovers all the metric
collectors in the assembly, and fetches all of the mertics they expose.
It then prints them out as a table that can be included in the Readme.